complete : entity

+++ src/main/java/com/danjam/users/Users.java
@@ -20,7 +20,7 @@ public class Users {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private long id;
+    private Long id;
 
     @Column(name = "email", nullable = false)
     private String email;
@@ -51,7 +51,7 @@ public class Users {
     private String status;
 
     @Builder
-    public Users(int id, String email, String password, String name, int phoneNum, Role role, LocalDateTime createAt, LocalDateTime updateAt, String status) {
+    public Users(Long id, String email, String password, String name, int phoneNum, Role role, LocalDateTime createAt, LocalDateTime updateAt, String status) {
         this.id = id;
         this.email = email;
         this.password = password;
